+++ shell/pwd.go
@@ -9,7 +9,10 @@ import (
 )
 
 // pwdCommand print current working directory.
-func (c CommandList) pwdCommand(_ context.Context, _ *Shell, _ []string) error {
+func (c CommandList) pwdCommand(_ context.Context, _ *Shell, argv []string) error {
+	if len(argv) > 0 {
+		return fmt.Errorf(".pwd takes no arguments, got %d", len(argv))
+	}
 	dir, err := os.Getwd()
 	if err != nil {
 		return err